# perf: rewrite skip function