--- a/pilota-build/test_data/thrift/decode_error.rs
+++ b/pilota-build/test_data/thrift/decode_error.rs
@@ -77,11 +77,13 @@ pub mod decode_error {
                 protocol.read_struct_end()?;
 
                 let Some(b) = b else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field b is required".to_string(),
-                    ));
-                };
+                            "field b is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { b };
                 Ok(data)
@@ -136,11 +138,13 @@ pub mod decode_error {
                 protocol.read_struct_end().await?;
 
                 let Some(b) = b else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field b is required".to_string(),
-                    ));
-                };
+                            "field b is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { b };
                 Ok(data)
@@ -230,11 +234,13 @@ pub mod decode_error {
                 protocol.read_struct_end()?;
 
                 let Some(a) = a else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field a is required".to_string(),
-                    ));
-                };
+                            "field a is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { a };
                 Ok(data)
@@ -289,11 +295,13 @@ pub mod decode_error {
                 protocol.read_struct_end().await?;
 
                 let Some(a) = a else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field a is required".to_string(),
-                    ));
-                };
+                            "field a is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { a };
                 Ok(data)
@@ -383,11 +391,13 @@ pub mod decode_error {
                 protocol.read_struct_end()?;
 
                 let Some(c) = c else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field c is required".to_string(),
-                    ));
-                };
+                            "field c is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { c };
                 Ok(data)
@@ -442,11 +452,13 @@ pub mod decode_error {
                 protocol.read_struct_end().await?;
 
                 let Some(c) = c else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field c is required".to_string(),
-                    ));
-                };
+                            "field c is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { c };
                 Ok(data)