--- a/pilota-build/test_data/thrift/normal.rs
+++ b/pilota-build/test_data/thrift/normal.rs
@@ -472,29 +472,37 @@ pub mod normal {
                 protocol.read_struct_end()?;
 
                 let Some(msg) = msg else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field msg is required".to_string(),
-                    ));
-                };
+                            "field msg is required".to_string()
+                    )
+                )
+            };
                 let Some(msg_map) = msg_map else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field msg_map is required".to_string(),
-                    ));
-                };
+                            "field msg_map is required".to_string()
+                    )
+                )
+            };
                 let Some(sub_msgs) = sub_msgs else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field sub_msgs is required".to_string(),
-                    ));
-                };
+                            "field sub_msgs is required".to_string()
+                    )
+                )
+            };
                 let Some(flag_msg) = flag_msg else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field flag_msg is required".to_string(),
-                    ));
-                };
+                            "field flag_msg is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self {
                     msg,
@@ -618,29 +626,37 @@ pub mod normal {
                 protocol.read_struct_end().await?;
 
                 let Some(msg) = msg else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field msg is required".to_string(),
-                    ));
-                };
+                            "field msg is required".to_string()
+                    )
+                )
+            };
                 let Some(msg_map) = msg_map else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field msg_map is required".to_string(),
-                    ));
-                };
+                            "field msg_map is required".to_string()
+                    )
+                )
+            };
                 let Some(sub_msgs) = sub_msgs else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field sub_msgs is required".to_string(),
-                    ));
-                };
+                            "field sub_msgs is required".to_string()
+                    )
+                )
+            };
                 let Some(flag_msg) = flag_msg else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field flag_msg is required".to_string(),
-                    ));
-                };
+                            "field flag_msg is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self {
                     msg,
@@ -913,11 +929,13 @@ pub mod normal {
                 protocol.read_struct_end()?;
 
                 let Some(req) = req else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field req is required".to_string(),
-                    ));
-                };
+                            "field req is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { req };
                 Ok(data)
@@ -976,11 +994,13 @@ pub mod normal {
                 protocol.read_struct_end().await?;
 
                 let Some(req) = req else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field req is required".to_string(),
-                    ));
-                };
+                            "field req is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { req };
                 Ok(data)
@@ -1598,11 +1618,13 @@ pub mod normal {
                 protocol.read_struct_end()?;
 
                 let Some(req) = req else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field req is required".to_string(),
-                    ));
-                };
+                            "field req is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { req };
                 Ok(data)
@@ -1661,11 +1683,13 @@ pub mod normal {
                 protocol.read_struct_end().await?;
 
                 let Some(req) = req else {
-                    return Err(::pilota::thrift::DecodeError::new(
+                return Err(
+                    ::pilota::thrift::DecodeError::new(
                         ::pilota::thrift::DecodeErrorKind::InvalidData,
-                        "field req is required".to_string(),
-                    ));
-                };
+                            "field req is required".to_string()
+                    )
+                )
+            };
 
                 let data = Self { req };
                 Ok(data)